services/clip-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
transformers==4.35.2
torch>=2.0.0
torchvision>=0.15.0
pillow==10.1.0
setuptools
numpy==1.24.3
faiss-cpu>=1.7.0
pydantic==2.5.0
python-dotenv==1.0.0
